Remove class NonAtomicCreateUnderFileSystem and temp creation logic to default UnderFileSystem
