+++ underfs/s3a/src/main/java/alluxio/underfs/s3a/S3AUnderFileSystem.java
@@ -238,6 +238,11 @@ public class S3AUnderFileSystem extends UnderFileSystem {
 
   @Override
   public OutputStream create(String path, CreateOptions options) throws IOException {
+    return createTemporary(path, options);
+  }
+
+  @Override
+  public OutputStream createTemporary(String path, CreateOptions options) throws IOException {
     if (mkdirs(getParentKey(path), true)) {
       // Return the direct stream if the user has enabled direct writes
       if (Configuration.getBoolean(PropertyKey.UNDERFS_S3A_DIRECT_WRITES_ENABLED)) {
